Trim common whitespace from included code snippets
When a set of lines are included into a DFM file, no indenting is
removed from the beginning of the line. This results in code snippets
that can potentially be shifted far to the right in the rendered
output.

This patch adds code that looks through the included lines for leading
spaces and tabs, and removes enough that any indent that is common to
all the lines is removed.