Change the format to json object instead
as for 'uri-reference', `#` means current document, however current document is the schema instead of the actual document, change it to `JSON-pointer` so that the value is actually the pointer to the actual document object, for example `#/definition/` references to the `definition` section defined in this schema. while we'd like `#/metadata` to point to the input object model.