Remove unused triple colon test.
The test in question was supposed to test nested blocks for triplecolon
only, but we don't have an extension that natively allows nesting (zones
cannot be nested and forms are self-closing blocks). The current test
used the moniker block, which is not part of the triple colon extension.
As such, even with an assertion the test would prove nothing.