Fix unit tests broken by upgrade of FCS
Somehow the FCS assembly includes a reference to netstandard2.0, which indirectly
references Microsoft.Build.Framework 4.0.0.0.  This isn't picked up by the toolchain
for some reason, and BRs aren't generated to redirect to the proper version.  BRs
already existed in the main docfx app.config, so we just copy those over to the
docfx.Tests project.