Don't warn for $schema element
Adding a `$schema` attribute pointing to the official schema is a nice way to get completion in an editor. This caused a warning during build, however.

This checks for that element when building the commands so we don't warn for what is already a supported scenario.

Fixes Using $schema to get completion causes a build warning #6926