Add strict-inline and note about SDP
PR for discussion, no need to merge if this is covered elsewhere.

In the marketing schema and stay current we've encountered the need to limit markdown to inline elements.

Also looking for the plan for how we'll reference the modes from SDP schemas. We have `contentType: markdown`. How do sanitization modes relate? `markdown-{sanitization-mode}`? `sanitizationMode` property?

@herohua @yufeih @wibjorn