Update index.md
I have updated the actions script because old version was not working properly in the newest versions of docfx. It now uses official upload pages actions instead of relying on the third party one. And now the default path is correct for website deployment: 'docfx_project/_site' instead of 'docs/_site'.