fix: use `textContent` to get header titles (#10790)
The modern template currently uses `innerText` to get the text from h2
and h3 tags for the "In This Article" section of the affix. However,
this can cause links with empty text content to be generated for headers
which are not visible at the moment the navigation is generated, such as
those in a `<details>` element. Using `textContent` instead will still
properly retrieve the text from these headers.

Co-authored-by: Preston Petrie <ppetrie@nyb.com>
Co-authored-by: Yufei Huang <yufeih@live.com>